moved test file and modified blas_ops.h commenting
